Adding final push to github

+++ moto-edge-service/src/main/java/com/trilogy/motoedgeservice/service/MotoService.java
@@ -19,7 +19,7 @@ public class MotoService {
         return client.getAll();
     }
 
-    public Motorcycle findOne(long id) {
+    public Motorcycle findOne(Integer id) {
         return client.findOne(id);
     }
